feat: support skipIf & runIf test API
